thirth

+++ books/resources/views/post/comment_list.blade.php
@@ -1,8 +1,10 @@
 @foreach ($review->comments as $comment) 
   <div class="mb-2">
+  @if (Auth::user() !== null)
     @if ($comment->user->id == Auth::user()->id)
       <a class="delete-comment" data-remote="true" rel="nofollow" data-method="delete" href="/comments/{{ $comment->id }}"></a>
     @endif
+  @endif
     <span>
       <strong>
         <a class="no-text-decoration black-color" href="/users/{{ $comment->user->id }}">{{ $comment->user->name }}</a>
